Prepare the basic structure
It looks like the Persistence part of the library will take too much of the time. Also being opionated on how the devices, transfers, and connections will be stored looks to be an unnecessary effort since it is better to keep it outside of this library which aims to be as generic as possible.