Implement key approval mechanism instead of blocking
When a device connected with a wrong key, we were blocking it from making further requests with the same key by actually blocking it, which were resulting in the blocking of the authentic device with the right key. Now, with this change, we create a request with new keys to be approved and check their existence when there is a new request.