protocol: Improve handling early exits & picture
Picture data was a part of the Client class, which
wasn't useful in a real use case. This change
removes that part and makes PersistenceProvider
responsible for it. Also, checksum is now is
considered 'revision number', so it doesn't mean
it is the hash code of the picture data but
rather a date on which the picture was changed.