encrpytion: Warn both parties on SSL errors
Previously, we were trying to warn the server
when there is an SSL error, indicating that a
client is pretending to be someone else. The
problem is that it is hard to accuretely to
detect which side is on the wrong using the
exception types because different JVM
implementations throw different types of
exceptions for the same type of error.

In the end, I decided to warn both parties.
because it is easier and should lead to 'less'
confusion.

Credentials should not change by default,
so I don't think any of the warning systems
will be in use that prominently anyway.